+++ .renvignore
@@ -16,3 +16,5 @@ __pycache__/
 
 # Ignore all subfolders
 */*
+
+alembic/
