+++ alembic/env.py
@@ -1,0 +1,3 @@
+from dsst_etl.models import Base 
+
+target_metadata = Base.metadata
