+++ dsst_etl/db.py
@@ -1,20 +1,11 @@
 import os
 
+from dsst_etl import get_db_engine
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 
 from .models import Base
 
-
-def get_db_engine():
-    database_url = (
-        "postgresql://"
-        f"{os.environ['POSTGRES_USER']}"
-        f":{os.environ['POSTGRES_PASSWORD']}"
-        f"@{os.environ['POSTGRES_HOST']}:"
-        f"{os.environ['POSTGRES_PORT']}"
-    )
-    return create_engine(database_url)
 
 
 def get_db_session():
